+++ packages/content/src/lib/markdown-content-renderer.service.ts
@@ -1,18 +1,12 @@
-import {
-  inject,
-  Injectable,
-  InjectionToken,
-  PLATFORM_ID,
-  Provider,
-} from '@angular/core';
+import { inject, Injectable, InjectionToken, Provider } from '@angular/core';
 import { getHeadingList } from 'marked-gfm-heading-id';
 
 import { ContentRenderer, TableOfContentItem } from './content-renderer';
 import { MarkedSetupService } from './marked-setup.service';
+import { RenderTaskService } from './render-task.service';
 
 @Injectable()
 export class MarkdownContentRendererService implements ContentRenderer {
-  platformId = inject(PLATFORM_ID);
   #marked = inject(MarkedSetupService, { self: true });
 
   async render(content: string): Promise<string> {
@@ -38,6 +32,8 @@ export function withMarkdownRenderer(
   options?: MarkdownRendererOptions
 ): Provider {
   return [
+    MarkedSetupService,
+    RenderTaskService,
     {
       provide: ContentRenderer,
       useFactory: () => new MarkdownContentRendererService(),
@@ -55,7 +51,7 @@ export function withMarkdownRenderer(
 }
 
 export function provideContent(...features: Provider[]) {
-  return [...features, MarkedSetupService];
+  return [...features];
 }
 
 export const MERMAID_IMPORT_TOKEN = new InjectionToken<