+++ packages/nx-plugin/src/generators/app/generator.spec.ts
@@ -94,8 +94,8 @@ describe('nx-plugin generator', () => {
     dependencies: Record<string, string>,
     devDependencies: Record<string, string>
   ) => {
-    expect(dependencies['@analogjs/content']).toBe('^1.0.0-rc.0');
-    expect(dependencies['@analogjs/router']).toBe('^1.0.0-rc.0');
+    expect(dependencies['@analogjs/content']).toBe('^1.2.0');
+    expect(dependencies['@analogjs/router']).toBe('^1.2.0');
     expect(dependencies['@angular/platform-server']).toBe(
       dependencies['@angular/core']
     );
@@ -111,10 +111,8 @@ describe('nx-plugin generator', () => {
     // we just check for truthy because @nx/eslint generator
     // will install the correct version based on Nx version
     expect(devDependencies['@nx/eslint']).toBeTruthy();
-    expect(devDependencies['@analogjs/platform']).toBe('^1.0.0-rc.0');
-    expect(devDependencies['@analogjs/vite-plugin-angular']).toBe(
-      '^1.0.0-rc.0'
-    );
+    expect(devDependencies['@analogjs/platform']).toBe('^1.2.0');
+    expect(devDependencies['@analogjs/vite-plugin-angular']).toBe('^1.2.0');
     expect(devDependencies['@nx/vite']).toBe('^17.0.0');
     expect(devDependencies['jsdom']).toBe('^22.0.0');
     expect(devDependencies['vite']).toBe('^4.4.8');
@@ -126,8 +124,8 @@ describe('nx-plugin generator', () => {
     dependencies: Record<string, string>,
     devDependencies: Record<string, string>
   ) => {
-    expect(dependencies['@analogjs/content']).toBe('^1.0.0-rc.0');
-    expect(dependencies['@analogjs/router']).toBe('^1.0.0-rc.0');
+    expect(dependencies['@analogjs/content']).toBe('^1.2.0');
+    expect(dependencies['@analogjs/router']).toBe('^1.2.0');
     expect(dependencies['@angular/platform-server']).toBe(
       dependencies['@angular/core']
     );
@@ -143,10 +141,8 @@ describe('nx-plugin generator', () => {
     // we just check for truthy because @nx/eslint generator
     // will install the correct version based on Nx version
     expect(devDependencies['@nx/eslint']).toBeTruthy();
-    expect(devDependencies['@analogjs/platform']).toBe('^1.0.0-rc.0');
-    expect(devDependencies['@analogjs/vite-plugin-angular']).toBe(
-      '^1.0.0-rc.0'
-    );
+    expect(devDependencies['@analogjs/platform']).toBe('^1.2.0');
+    expect(devDependencies['@analogjs/vite-plugin-angular']).toBe('^1.2.0');
     expect(devDependencies['@nx/vite']).toBe('^18.0.0');
     expect(devDependencies['jsdom']).toBe('^22.1.0');
     expect(devDependencies['vite']).toBe('^5.0.0');